+++ src/main/kotlin/nl/jrdie/idea/springql/index/entry/QLMethodBatchMappingIndexEntry.kt
@@ -1,15 +1,19 @@
 package nl.jrdie.idea.springql.index.entry
 
 import com.intellij.psi.PsiElement
+import org.jetbrains.uast.UAnnotation
+import org.jetbrains.uast.UMethod
 
 data class QLMethodBatchMappingIndexEntry(
-    val parentType: String?,
-    val field: String?,
-    val annotationPsi: PsiElement,
+    override val parentType: String?,
+    override val field: String?,
+    override val annotationPsi: PsiElement,
     val methodPsi: PsiElement,
     /**
      * @see
      */
     val conformsToReturnSpec: Boolean,
-    val schemaPsi: List<PsiElement>
-) : QLIndexEntry
+    override val schemaPsi: List<PsiElement>,
+    override val uAnnotation: UAnnotation,
+    val uMethod: UMethod
+) : SchemaMappingIndexEntry