+++ src/main/kotlin/nl/jrdie/idea/springql/index/entry/QLMethodSchemaMappingIndexEntry.kt
@@ -2,6 +2,7 @@ package nl.jrdie.idea.springql.index.entry
 
 import com.intellij.psi.PsiElement
 import org.jetbrains.uast.UAnnotation
+import org.jetbrains.uast.UMethod
 
 data class QLMethodSchemaMappingIndexEntry(
     val methodPsi: PsiElement,
@@ -9,5 +10,6 @@ data class QLMethodSchemaMappingIndexEntry(
     override val field: String?,
     override val annotationPsi: PsiElement,
     override val schemaPsi: List<PsiElement>,
-    override val uAnnotation: UAnnotation
+    override val uAnnotation: UAnnotation,
+    val uMethod: UMethod,
 ) : SchemaMappingIndexEntry, QLIndexEntry