+++ src/main/kotlin/nl/jrdie/idea/springql/types/SchemaMappingSummary.kt
@@ -2,6 +2,7 @@ package nl.jrdie.idea.springql.types
 
 import com.intellij.psi.PsiElement
 import org.jetbrains.uast.UAnnotation
+import org.jetbrains.uast.UMethod
 
 data class SchemaMappingSummary(
     val typeName: String,
@@ -9,5 +10,8 @@ data class SchemaMappingSummary(
     val annotationPsi: PsiElement,
     val schemaPsi: PsiElement?,
     val uAnnotation: UAnnotation,
-    val annotationName: String
+    val annotationName: String,
+    val uMethod: UMethod,
+    val mappingType: SchemaMappingType,
+    val methodPsi: PsiElement
 )