+++ src/main/kotlin/nl/jrdie/idea/springql/types/SchemaMappingType.kt
@@ -1,0 +1,6 @@
+package nl.jrdie.idea.springql.types
+
+enum class SchemaMappingType {
+    SCHEMA_MAPPING,
+    BATCH_MAPPING
+}